% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Malaria-data.R
\docType{data}
\name{Malaria}
\alias{Malaria}
\alias{Malaria2}
\title{Malaria}
\format{


An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 2 columns.
}
\source{
\url{https://www.nih.gov/news-events/news-releases/experimental-pfspz-malaria-vaccine-provides-durable-protection-against-multiple-strains-nih-clinical-trial}
}
\usage{
data(Malaria)

data(Malaria2)
}
\description{
Results of a small Malaria vaccine trial in which 14 subjects were given a
new vaccine (PfSPZ) and 6 were given a placebo. After 19 weeks, both groups
were exposed to a
drug-sensitive malaria virus strain (allowing any infections
to be treated effectively).
}
\details{
The two variables are
\itemize{
\item \code{group} -- A factor with levels Treatment and Placebo
\item \code{malaria} -- A factor with levels Yes and No, indicating whether the subject
exhibited malaria symptoms
}

In \code{Malaria2} these are abbreviated to a single letter.
}
\examples{
require(mosaic)
Malaria \%>\%
  dplyr::group_by(group, malaria, .drop = FALSE) \%>\%
  dplyr::tally()

tally(malaria ~ group, data = Malaria)
MalariaSims <- do(1000) * diffprop(malaria ~ shuffle(group), data = Malaria)
MalariaSims \%>\%
  gf_histogram(~diffprop)

mosaic::tally(malaria ~ group, data = Malaria2)
}
\keyword{datasets}
